@@ -11,7 +11,7 @@
     }
   ],
   "require": {
-    "php": "^7.4|^8",
+    "php": "^8",
     "ext-sockets": "*"
   },
   "license": "MIT",
